+++ src/components/AdminInterface/Board/Notice/NoticeDetail.jsx
